feat: converted header and footer to tsx
