+++ src/prefabs/pageWithHeaderAndFooter.tsx
@@ -1,0 +1,518 @@
+import * as React from 'react';
+import {
+  prefab as makePrefab,
+  Icon,
+  option,
+  text,
+  sizes,
+  size,
+  showIf,
+  toggle,
+  color,
+  ThemeColor,
+  PrefabReference,
+  BeforeCreateArgs,
+  variable,
+  font,
+  showIfTrue,
+} from '@betty-blocks/component-sdk';
+import { Box as prefabBox } from './structures/Box';
+import { options as boxOptions } from './structures/Box/options';
+import { Column } from './structures/Column';
+import { options as columnOptions } from './structures/Column/options';
+import { Grid } from './structures/Grid';
+import { options as gridOptions } from './structures/Grid/options';
+import { Row } from './structures/Row';
+import { options as rowOptions } from './structures/Row/options';
+import { Text as TextPrefab } from './structures/Text';
+import { options as textOptions } from './structures/Text/options';
+import { appBar } from './structures/Appbar';
+import { options as appBarOptions } from './structures/Appbar/options';
+import { OpenPageButton } from './structures/OpenPage';
+import { options as openPageOptions } from './structures/Button/options';
+
+const attrs = {
+  name: 'Header and footer',
+  icon: Icon.NavbarIcon,
+  type: 'page',
+  description: 'Full height page with a header and footer',
+  detail:
+    'Start with a full height page containing a header, footer and body that will scale to the height of the content.',
+  previewUrl: 'https://preview.betty.app/header-and-footer',
+  previewImage:
+    'https://assets.bettyblocks.com/efaf005f4d3041e5bdfdd0643d1f190d_assets/files/Page_Template_Header_And_Footer.jpg',
+  category: 'LAYOUT',
+};
+
+const beforeCreate = ({
+  save,
+  close,
+  prefab,
+  components: { Header, Content, PartialSelector, Footer, Field, Box, Text },
+}: BeforeCreateArgs) => {
+  const [headerPartialId, setHeaderPartialId] = React.useState('');
+  const [footerPartialId, setFooterPartialId] = React.useState('');
+
+  const getDescendantByRef = (refValue: string, structure: any) =>
+    structure.reduce((acc: string, component: PrefabReference) => {
+      if (acc) return acc;
+      if (
+        component.type === 'COMPONENT' &&
+        // eslint-disable-next-line no-prototype-builtins
+        component.ref
+          ? Object.values(component.ref).indexOf(refValue) > -1
+          : undefined
+      ) {
+        return component;
+      }
+      if (component.type === 'PARTIAL') {
+        return acc;
+      }
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      return getDescendantByRef(refValue, component.descendants);
+    }, null);
+
+  return (
+    <>
+      <Header onClose={close} title="Configure header and footer" />
+
+      <Content>
+        <Box pad={{ bottom: '15px' }}>
+          <Box>
+            <Text size="medium" weight="bolder">
+              Select partials
+            </Text>
+          </Box>
+          <Box pad={{ bottom: '15px' }}>
+            <Text color="grey700">
+              By using a partial for the top menu and footer you can easily
+              reuse the same structure without having to go through every page.
+            </Text>
+          </Box>
+          <Field label="TOP MENU PARTIAL">
+            <PartialSelector
+              label="Select a partial"
+              onChange={(headerId: string) => {
+                setHeaderPartialId(headerId);
+              }}
+              preSelected="Top menu"
+              value={headerPartialId}
+              allowedTypes={[
+                'BODY_COMPONENT',
+                'CONTAINER_COMPONENT',
+                'CONTENT_COMPONENT',
+              ]}
+            />
+          </Field>
+        </Box>
+        <Box pad={{ bottom: '15px' }}>
+          <Field label="FOOTER PARTIAL">
+            <PartialSelector
+              label="Select a partial"
+              onChange={(footerId: string) => {
+                setFooterPartialId(footerId);
+              }}
+              preSelected="Footer"
+              value={footerPartialId}
+              allowedTypes={[
+                'BODY_COMPONENT',
+                'CONTAINER_COMPONENT',
+                'CONTENT_COMPONENT',
+              ]}
+            />
+          </Field>
+        </Box>
+      </Content>
+      <Footer
+        onClick={close}
+        onSave={() => {
+          const newPrefab = { ...prefab };
+          const prefabFooter = getDescendantByRef(
+            '#Footer',
+            newPrefab.structure,
+          );
+          const prefabHeader = getDescendantByRef(
+            '#Header',
+            newPrefab.structure,
+          );
+          if (headerPartialId) {
+            prefabHeader.descendants = [{ type: 'PARTIAL', partialId: '' }];
+            prefabHeader.descendants[0].partialId = headerPartialId;
+          }
+
+          if (footerPartialId) {
+            prefabFooter.descendants = [{ type: 'PARTIAL', partialId: '' }];
+            prefabFooter.descendants[0].partialId = footerPartialId;
+          }
+
+          save(newPrefab);
+        }}
+        onSkip={() => {
+          const newPrefab = { ...prefab };
+          save(newPrefab);
+        }}
+      />
+    </>
+  );
+};
+
+export default makePrefab('TSX Header & Footer', attrs, beforeCreate, [
+  Row(
+    {
+      options: {
+        ...rowOptions,
+        maxRowWidth: option('CUSTOM', {
+          label: 'Width',
+          value: 'Full',
+          configuration: {
+            as: 'BUTTONGROUP',
+            dataType: 'string',
+            allowedInput: [
+              { name: 'S', value: 'S' },
+              { name: 'M', value: 'M' },
+              { name: 'L', value: 'L' },
+              { name: 'XL', value: 'XL' },
+              { name: 'Full', value: 'Full' },
+            ],
+          },
+        }),
+        rowHeight: text('Height', {
+          value: '100%',
+          configuration: {
+            as: 'UNIT',
+          },
+        }),
+      },
+    },
+    [
+      Column(
+        {
+          options: {
+            ...columnOptions,
+            columnHeight: text('Height', {
+              value: '100%',
+              configuration: {
+                as: 'UNIT',
+              },
+            }),
+            innerSpacing: sizes('Inner space', {
+              value: ['0rem', '0rem', '0rem', '0rem'],
+            }),
+          },
+        },
+        [
+          Grid(
+            {
+              options: {
+                ...gridOptions,
+                height: size('Height', {
+                  value: '100%',
+                  configuration: {
+                    as: 'UNIT',
+                  },
+                }),
+              },
+            },
+            [
+              Grid(
+                {
+                  options: {
+                    ...gridOptions,
+                    direction: option('CUSTOM', {
+                      value: 'column',
+                      label: 'Direction',
+                      configuration: {
+                        as: 'BUTTONGROUP',
+                        dataType: 'string',
+                        allowedInput: [
+                          { name: 'Horizontal', value: 'row' },
+                          { name: 'Vertical', value: 'column' },
+                        ],
+                        condition: showIf('type', 'EQ', 'container'),
+                      },
+                    }),
+                  },
+                },
+                [
+                  prefabBox(
+                    {
+                      ref: { id: '#Header' },
+
+                      options: {
+                        ...boxOptions,
+                        innerSpacing: sizes('Inner space', {
+                          value: ['0rem', '0rem', '0rem', '0rem'],
+                        }),
+                      },
+                    },
+                    [
+                      Column(
+                        {
+                          options: {
+                            ...columnOptions,
+                            innerSpacing: sizes('Inner space', {
+                              value: ['0rem', '0rem', '0rem', '0rem'],
+                            }),
+                          },
+                        },
+                        [
+                          prefabBox(
+                            {
+                              options: {
+                                ...boxOptions,
+                                innerSpacing: sizes('Inner space', {
+                                  value: ['0rem', '0rem', '0rem', '0rem'],
+                                }),
+                                backgroundOptions: toggle(
+                                  'Show background options',
+                                  {
+                                    value: true,
+                                  },
+                                ),
+                                backgroundColor: color('Background color', {
+                                  value: ThemeColor.PRIMARY,
+                                  configuration: {
+                                    condition: showIf(
+                                      'backgroundOptions',
+                                      'EQ',
+                                      true,
+                                    ),
+                                  },
+                                }),
+                              },
+                            },
+                            [
+                              Row({}, [
+                                Column(
+                                  {
+                                    options: {
+                                      ...columnOptions,
+                                      innerSpacing: sizes('Inner space', {
+                                        value: ['0rem', '0rem', '0rem', '0rem'],
+                                      }),
+                                    },
+                                  },
+                                  [
+                                    appBar(
+                                      {
+                                        options: {
+                                          ...appBarOptions,
+                                          logoSource: variable('Logo', {
+                                            value: [
+                                              'https://assets.bettyblocks.com/efaf005f4d3041e5bdfdd0643d1f190d_assets/files/Your_Logo_-_W.svg',
+                                            ],
+                                          }),
+                                          title: variable('Title', {
+                                            value: [],
+                                          }),
+                                        },
+                                      },
+                                      [
+                                        OpenPageButton(
+                                          {
+                                            style: {
+                                              overwrite: {
+                                                backgroundColor: {
+                                                  type: 'STATIC',
+                                                  value: 'transparent',
+                                                },
+                                                boxShadow: 'none',
+                                                color: {
+                                                  type: 'THEME_COLOR',
+                                                  value: 'white',
+                                                },
+                                                fontFamily: 'Roboto',
+                                                fontSize: '0.875rem',
+                                                fontStyle: 'none',
+                                                fontWeight: '400',
+                                                padding: ['0rem', '0rem'],
+                                                textDecoration: 'none',
+                                                textTransform: 'none',
+                                              },
+                                            },
+                                            options: {
+                                              ...openPageOptions,
+                                              buttonText: variable(
+                                                'Button text',
+                                                { value: ['Menu 1'] },
+                                              ),
+                                              outerSpacing: sizes(
+                                                'Outer space',
+                                                {
+                                                  value: [
+                                                    '0rem',
+                                                    'M',
+                                                    '0rem',
+                                                    'M',
+                                                  ],
+                                                },
+                                              ),
+                                            },
+                                          },
+                                          [],
+                                        ),
+                                        OpenPageButton(
+                                          {
+                                            style: {
+                                              overwrite: {
+                                                backgroundColor: {
+                                                  type: 'STATIC',
+                                                  value: 'transparent',
+                                                },
+                                                boxShadow: 'none',
+                                                color: {
+                                                  type: 'THEME_COLOR',
+                                                  value: 'white',
+                                                },
+                                                fontFamily: 'Roboto',
+                                                fontSize: '0.875rem',
+                                                fontStyle: 'none',
+                                                fontWeight: '400',
+                                                padding: ['0rem', '0rem'],
+                                                textDecoration: 'none',
+                                                textTransform: 'none',
+                                              },
+                                            },
+                                            options: {
+                                              ...openPageOptions,
+                                              buttonText: variable(
+                                                'Button text',
+                                                { value: ['Menu 2'] },
+                                              ),
+                                              outerSpacing: sizes(
+                                                'Outer space',
+                                                {
+                                                  value: [
+                                                    '0rem',
+                                                    'M',
+                                                    '0rem',
+                                                    '0rem',
+                                                  ],
+                                                },
+                                              ),
+                                            },
+                                          },
+                                          [],
+                                        ),
+                                      ],
+                                    ),
+                                  ],
+                                ),
+                              ]),
+                            ],
+                          ),
+                        ],
+                      ),
+                    ],
+                  ),
+                  prefabBox(
+                    {
+                      options: {
+                        ...boxOptions,
+                        stretch: toggle('Stretch (when in flex container)', {
+                          value: true,
+                        }),
+                        width: size('Width', {
+                          value: '100%',
+                          configuration: {
+                            as: 'UNIT',
+                          },
+                        }),
+                        innerSpacing: sizes('Inner space', {
+                          value: ['0rem', '0rem', '0rem', '0rem'],
+                        }),
+                        backgroundColor: color('Background color', {
+                          value: ThemeColor.LIGHT,
+                          configuration: {
+                            condition: showIf('backgroundOptions', 'EQ', true),
+                          },
+                        }),
+                        backgroundColorAlpha: option('NUMBER', {
+                          label: 'Background color opacity',
+                          value: 20,
+                          configuration: {
+                            condition: showIf('backgroundOptions', 'EQ', true),
+                          },
+                        }),
+                      },
+                    },
+                    [],
+                  ),
+                  prefabBox(
+                    {
+                      ref: { id: '#Footer' },
+                      options: {
+                        ...boxOptions,
+                        width: size('Width', {
+                          value: '100%',
+                          configuration: {
+                            as: 'UNIT',
+                          },
+                        }),
+                        innerSpacing: sizes('Inner space', {
+                          ref: { id: '#footerBoxInnerSpacing' },
+                          value: ['0rem', '0rem', '0rem', '0rem'],
+                        }),
+                        backgroundColor: color('Background color', {
+                          value: ThemeColor.LIGHT,
+                          configuration: {
+                            condition: showIf('backgroundOptions', 'EQ', true),
+                          },
+                        }),
+                        backgroundColorAlpha: option('NUMBER', {
+                          label: 'Background color opacity',
+                          value: 20,
+                          configuration: {
+                            condition: showIf('backgroundOptions', 'EQ', true),
+                          },
+                        }),
+                      },
+                    },
+                    [
+                      TextPrefab(
+                        {
+                          options: {
+                            ...textOptions,
+                            content: variable('Content', {
+                              value: ['Powered by Bettyblocks'],
+                              configuration: { as: 'MULTILINE' },
+                            }),
+                            textAlignment: option('CUSTOM', {
+                              label: 'Text Alignment',
+                              value: 'center',
+                              configuration: {
+                                as: 'BUTTONGROUP',
+                                dataType: 'string',
+                                allowedInput: [
+                                  { name: 'Left', value: 'left' },
+                                  { name: 'Center', value: 'center' },
+                                  { name: 'Right', value: 'right' },
+                                ],
+                              },
+                            }),
+                            type: font('Font', { value: ['Body1'] }),
+                            outerSpacing: sizes('Outer space', {
+                              value: ['L', 'L', 'L', 'L'],
+                            }),
+                            styles: toggle('Styles', { value: false }),
+                            textColor: color('Text color', {
+                              value: ThemeColor.MEDIUM,
+                              configuration: {
+                                condition: showIfTrue('styles'),
+                              },
+                            }),
+                          },
+                        },
+                        [],
+                      ),
+                    ],
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ],
+      ),
+    ],
+  ),
+]);
